refactor: resolve PR review blockers by consolidating helpers and finalizing DoctorService extraction
