+++ PoshMcp.Server/Configuration/ConfigurationHelpers.cs
@@ -1,0 +1,125 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Text.RegularExpressions;
+using ModelContextProtocol.Server;
+
+namespace PoshMcp;
+
+internal static class ConfigurationHelpers
+{
+    internal static string DescribeConfigurationPath(string? configurationPath)
+    {
+        return string.IsNullOrWhiteSpace(configurationPath)
+            ? "(environment-only configuration)"
+            : configurationPath;
+    }
+
+    internal static string ToToolName(string functionName)
+    {
+        var normalized = functionName.Replace('-', '_');
+        normalized = Regex.Replace(normalized, "([a-z0-9])([A-Z])", "$1_$2");
+        normalized = Regex.Replace(normalized, "_+", "_");
+        return normalized.ToLowerInvariant();
+    }
+
+    internal static List<string> GetDiscoveredToolNames(List<McpServerTool> tools)
+    {
+        var names = new List<string>();
+
+        foreach (var tool in tools)
+        {
+            var name = TryGetNameFromObject(tool, 0);
+            if (!string.IsNullOrWhiteSpace(name))
+            {
+                names.Add(name);
+            }
+        }
+
+        return names
+            .Where(name => !string.IsNullOrWhiteSpace(name))
+            .Distinct(StringComparer.OrdinalIgnoreCase)
+            .OrderBy(name => name, StringComparer.OrdinalIgnoreCase)
+            .ToList();
+    }
+
+    internal static List<string> GetExpectedToolNames<TConfiguredFunctionStatus>(
+        IEnumerable<TConfiguredFunctionStatus> configuredFunctionStatus,
+        Func<TConfiguredFunctionStatus, IEnumerable<string>> matchedToolNamesSelector,
+        bool enableDynamicReloadTools)
+    {
+        var names = new List<string>();
+
+        names.AddRange(configuredFunctionStatus.SelectMany(matchedToolNamesSelector));
+
+        names.Add("get_last_command_output");
+        names.Add("sort_last_command_output");
+        names.Add("filter_last_command_output");
+        names.Add("group_last_command_output");
+
+        if (enableDynamicReloadTools)
+        {
+            names.Add("reload_configuration_from_file");
+            names.Add("update_configuration");
+            names.Add("get_configuration_status");
+        }
+
+        return names
+            .Distinct(StringComparer.OrdinalIgnoreCase)
+            .OrderBy(name => name, StringComparer.OrdinalIgnoreCase)
+            .ToList();
+    }
+
+    private static string? TryGetNameFromObject(object? value, int depth)
+    {
+        if (value is null || depth > 3)
+        {
+            return null;
+        }
+
+        var type = value.GetType();
+        var flags = BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic;
+
+        var directNameProperty = type.GetProperty("Name", flags);
+        if (directNameProperty is not null && directNameProperty.PropertyType == typeof(string))
+        {
+            var name = directNameProperty.GetValue(value) as string;
+            if (!string.IsNullOrWhiteSpace(name))
+            {
+                return name;
+            }
+        }
+
+        foreach (var property in type.GetProperties(flags))
+        {
+            if (property.GetIndexParameters().Length > 0)
+            {
+                continue;
+            }
+
+            object? nestedValue;
+            try
+            {
+                nestedValue = property.GetValue(value);
+            }
+            catch
+            {
+                continue;
+            }
+
+            if (nestedValue is null || nestedValue is string)
+            {
+                continue;
+            }
+
+            var nestedName = TryGetNameFromObject(nestedValue, depth + 1);
+            if (!string.IsNullOrWhiteSpace(nestedName))
+            {
+                return nestedName;
+            }
+        }
+
+        return null;
+    }
+}
