+++ PoshMcp.Server/Diagnostics/ConfiguredFunctionStatus.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace PoshMcp;
+
+public sealed record ConfiguredFunctionStatus(
+    string FunctionName,
+    string ExpectedToolName,
+    bool Found,
+    List<string> MatchedToolNames,
+    string? ResolutionReason = null);
